% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alpha_diversity.R
\name{get_alpha_diversity}
\alias{get_alpha_diversity}
\title{Calculates a alpha diversity metric}
\usage{
get_alpha_diversity(tbl, index = "shannon", ...)
}
\arguments{
\item{tbl}{abundance tibble with column sample_id. Denoised raw counts preferred.}

\item{index}{alpha diversity method}
}
\value{
alpha diversity tibble
}
\description{
Calculates a alpha diversity metric
}
