% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alpha_diversity.R
\name{get_alpha_diversity_all}
\alias{get_alpha_diversity_all}
\title{Calculates all alpha diversity metrics}
\usage{
get_alpha_diversity_all(tbl, ...)
}
\arguments{
\item{tbl}{abundance tibble with column sample_id. Denoised raw counts preferred.}
}
\value{
alpha diversity tibble with columns sample_id, shannon, simpson, invsimpson
}
\description{
Calculates all alpha diversity metrics
}
