% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_correlation_graph.R
\name{get_correlation_graph}
\alias{get_correlation_graph}
\title{Get correlation graph and calculate topology properties}
\usage{
get_correlation_graph(edges_tbl, nodes_tbl)
}
\arguments{
\item{edges_tbl}{edge tibble with columns from and to}

\item{nodes_tbl}{node tibble with column name}
}
\description{
Get correlation graph and calculate topology properties
}
