% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups_cor_tbl.R
\name{get_groups_cor_tbl}
\alias{get_groups_cor_tbl}
\title{Create nested correlation tibble containing plots and graphs}
\usage{
get_groups_cor_tbl(
  correlation_dir,
  nodes_tbl,
  features_grouping_col = "order",
  n_features = 8,
  max_p_value = 0.05,
  min_abs_cor = 0.1
)
}
\arguments{
\item{correlation_dir}{Path to correlation directory of DAnIEL. Must contain file groups.txt and a subdirectiry for each group}
}
\description{
Create nested correlation tibble containing plots and graphs
}
