% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_selected_params.R
\name{get_selected_params}
\alias{get_selected_params}
\title{Get selected parameters for all steps based on project$selected_analysis_params}
\usage{
get_selected_params(project)
}
\arguments{
\item{project}{list of DAnIEL file project.json}
}
\description{
Get selected parameters for all steps based on project$selected_analysis_params
}
