% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alpha_diversity.R
\name{plot_alpha_diversity}
\alias{plot_alpha_diversity}
\title{Plot allpha diversity box plot}
\usage{
plot_alpha_diversity(
  abundance_tbl,
  samples_tbl = NULL,
  grouping = NULL,
  singnif_bars = TRUE
)
}
\arguments{
\item{abundance_tbl}{abudnance tibble with colum sample_id. Raw denoised counts preferred.}

\item{samples_tbl}{samples meta data tibble with column sample_id. Only needed if grouping is provided}

\item{grouping}{name of column in samples_tbl used to group samples}

\item{singnif_bars}{show significance bars (Kruskal_Wallis and post hoc Dunn's test)}
}
\value{
ggplot boxplot of alpha diversities
}
\description{
Plot allpha diversity box plot
}
