% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation_topology.R
\name{plot_correlation_topology}
\alias{plot_correlation_topology}
\title{Generates boxplot of network topology properties}
\usage{
plot_correlation_topology(node_topology_tbl)
}
\arguments{
\item{node_topology_tbl}{tibble about network topology with columns name, correlation_group, key, and value}
}
\description{
Generates boxplot of network topology properties
}
