% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{plot_roc}
\alias{plot_roc}
\title{Plots ROC over resamples}
\usage{
plot_roc(fs, folds = FALSE)
}
\arguments{
\item{folds}{show ROC of all resamples as well}

\item{model}{feature selection object (e.g. caret sbf, rfe. train works as well)}
}
\description{
Plots ROC over resamples
}
