% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_hoc_box_plot.R
\name{post_hoc_box_plot}
\alias{post_hoc_box_plot}
\title{Post hoc box plot}
\usage{
post_hoc_box_plot(
  feature_tbl,
  post_stat_results_tbl,
  current_feature,
  group_var
)
}
\arguments{
\item{feature_tbl}{featute table}

\item{post_stat_results_tbl}{statistics results table of post doc tests}

\item{current_feature}{current feature used to filter feature_tbl}

\item{group_var}{grouping variable aka x axis of this plot}
}
\value{
Post hoc box plot (ggplot2)
}
\description{
Post hoc box plot
}
