% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quietly.R
\name{quietly}
\alias{quietly}
\title{supress any output from a function}
\usage{
quietly(x)
}
\description{
supress any output from a function
}
\seealso{
https://r.789695.n4.nabble.com/Suppressing-output-e-g-from-cat-td859876.html
}
