% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_guess_coltypes.R
\name{read_csv_guess_coltypes}
\alias{read_csv_guess_coltypes}
\title{Reads csv file and guessses column classes}
\usage{
read_csv_guess_coltypes(csv_path)
}
\arguments{
\item{csv_path}{path of csv file}
}
\value{
tibble with
}
\description{
Reads csv file and guessses column classes
}
