% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Create analysis report}
\usage{
report(
  project_dir,
  out_html,
  template_rmd = system.file("extdata", "report.Rmd", package = "danielLib")
)
}
\arguments{
\item{project_dir}{path of project directory the report shoulb be based on}

\item{out_html}{path to output report html file}
}
\description{
The report includes some chunks only if the corresponding analysis method was performed
This is done using conditional R markdown chunks with the asis engine (as is)
This engine, however, obviously do not support the execution of R code inside that chunk
Therefore, dynamic variables inside these chunks are reported as placeholders {{name}}
This placeholders have to be replaced in post production
}
