% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_compare_medians.R
\name{stat_compare_medians}
\alias{stat_compare_medians}
\title{Add median comparison P-values to a ggplot}
\usage{
stat_compare_medians(
  data = NULL,
  g = NULL,
  x = NULL,
  facet_key = NULL,
  facet_val = NULL,
  max_pvalue = 0.05,
  signif_bar_y_spacing = 0.1,
  ...
)
}
\arguments{
\item{data}{data used by ggplot}

\item{g}{name of grouping column used for dunn.test}

\item{x}{name of value column used for dunn.test}

\item{max_pvalue}{maximum adjusted p value of by dunn.test to display}

\item{signif_bar_y_spacing}{fraction of plot hight used as spacer between two significance bars}

\item{facet}{column of data used for facetting}
}
\description{
Add median comparison P-values to a ggplot
}
